+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## 0.1.7 -- UNRELEASED
 
+Changed the timer macro to no longer invoke timer*; timer* is now deprecated.
+
 ## 0.1.6 -- 10 Feb 2014
 
 Quick update to convert checkpoint* to a macro, so that it can be used inside a clojure.core.async go